# api/openapi.yaml
openapi: 3.1.0
info:
  title: Mesaar Logistics Operations API
  version: "1.0.0-draft"
  description: >
    Phase 2 contract draft. Reflects the EXISTING routes in `app/api/routes/*`
    remounted under the `/v1` prefix (ADR-005), plus clearly-tagged
    `driver-self (planned)` endpoints the delivered driver app needs
    (see docs/api-gap-analysis.md). Schemas mirror `app/schemas/*` and
    `app/models/*`. This file is hand-curated and is diffed in CI against the
    FastAPI-generated schema.
servers:
  - url: /v1
    description: Versioned base path
security:
  - bearerAuth: []
...
paths:
...
  /warehouses:
    post:
      tags: [warehouses]
      summary: Create a warehouse (admin, manager).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WarehouseCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WarehouseRead' }
        '400': { $ref: '#/components/responses/BadRequest' }
    get:
      tags: [warehouses]
      summary: List warehouses (admin, manager).
      parameters:
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
        - { name: limit, in: query, schema: { type: integer, default: 100 } }
      responses:
        '200':
          description: Page of warehouses
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/WarehouseRead' }
  /warehouses/{warehouse_id}:
    parameters:
      - { name: warehouse_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [warehouses]
      summary: Get a warehouse by ID (admin, manager).
      responses:
        '200':
          description: Warehouse
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WarehouseRead' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [warehouses]
      summary: Update a warehouse (admin, manager).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WarehouseUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WarehouseRead' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [warehouses]
      summary: Delete a warehouse (admin, manager).
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    BadRequest:
      description: Validation/business-rule error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      properties:
        detail: { type: string }
      required: [detail]
...
    WarehouseBase:
      type: object
      required: [code, name, address_line1, city, country, capacity_weight_kg, capacity_volume_m3]
      properties:
        code: { type: string, minLength: 2, maxLength: 32 }
        name: { type: string, minLength: 2, maxLength: 255 }
        address_line1: { type: string, maxLength: 255 }
        address_line2: { type: string, maxLength: 255, nullable: true }
        city: { type: string, maxLength: 128 }
        state: { type: string, maxLength: 128, nullable: true }
        country: { type: string, maxLength: 128 }
        postal_code: { type: string, maxLength: 32, nullable: true }
        latitude: { type: number, minimum: -90, maximum: 90, nullable: true }
        longitude: { type: number, minimum: -180, maximum: 180, nullable: true }
        capacity_weight_kg: { type: number, exclusiveMinimum: 0 }
        capacity_volume_m3: { type: number, exclusiveMinimum: 0 }
        max_daily_shipments: { type: integer, minimum: 1, nullable: true }
    WarehouseCreate:
      $ref: '#/components/schemas/WarehouseBase'
    WarehouseRead:
      allOf:
        - $ref: '#/components/schemas/WarehouseBase'
        - type: object
          required: [id, created_at, updated_at]
          properties:
            id: { type: string, format: uuid }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    WarehouseUpdate:
      type: object
      properties:
        name: { type: string, maxLength: 255 }
        address_line1: { type: string, maxLength: 255 }
        address_line2: { type: string, maxLength: 255, nullable: true }
        city: { type: string, maxLength: 128 }
        state: { type: string, maxLength: 128, nullable: true }
        country: { type: string, maxLength: 128 }
        postal_code: { type: string, maxLength: 32, nullable: true }
        latitude: { type: number, minimum: -90, maximum: 90, nullable: true }
        longitude: { type: number, minimum: -180, maximum: 180, nullable: true }
        capacity_weight_kg: { type: number, exclusiveMinimum: 0 }
        capacity_volume_m3: { type: number, exclusiveMinimum: 0 }
        max_daily_shipments: { type: integer, minimum: 1, nullable: true }
...